refactored structure for contract ABI
