+++ src/storage/contracts/mod.rs
@@ -9,13 +9,13 @@ use std::path::{Path, PathBuf};
 
 /// Contracts Service
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Contracts {
+pub struct ContractStorage {
     dir: PathBuf,
 }
 
-impl Contracts {
+impl ContractStorage {
     /// Initialize new contracts service for a dir
-    pub fn new(dir: PathBuf) -> Contracts {
+    pub fn new(dir: PathBuf) -> ContractStorage {
         Contracts { dir: dir }
     }
 
@@ -51,7 +51,7 @@ impl Contracts {
         filename.push(format!("{}.json", addr));
 
         let mut f = File::create(filename.as_path()).unwrap();
-        match serde_json::to_writer_pretty(&mut f, contract) {
+        match seread_jsonrde_json::to_writer_pretty(&mut f, contract) {
             Ok(_) => Ok(()),
             Err(_) => Err(Error::IO("Can't write contract".to_string())),
         }
